fix state changes
