+++ clients/fj-svelte/src/App.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
   import Client from "./lib/Client.svelte";
   import Loading from "./lib/Loading.svelte";
-  import { needs_questions, type CompTimes, type QuestionMeta } from "./utils";
+  import { CompState, get_current_comp_state, get_time_till_next_state, needs_questions, type CompTimes, type QuestionMeta } from "./utils";
 
   import { get_questions, get_comp_info, get_comp_times } from "./api";
+    import { onDestroy } from "svelte";
 
   let comp_times: CompTimes | undefined = undefined;
   let questions: Record<string, QuestionMeta> | undefined = undefined;
@@ -23,15 +24,36 @@
     });
   }
 
+  let current_state: CompState | undefined = undefined;
+  let timer_ref = -1;
+  // Sets a timer to trigger a rerender to update the main content
+  const set_timer_for_next_state = () => {
+    if(comp_times == undefined) return;
+    current_state = get_current_comp_state(comp_times);
+    if(needs_questions(comp_times)) {
+      load_questions();
+    }
+    console.log(`changed current state to ${current_state}`);
+
+    timer_ref = setTimeout(() => {
+      set_timer_for_next_state();
+    }, get_time_till_next_state(comp_times, current_state));
+  };
+  onDestroy(() => clearTimeout(timer_ref))
+  
+  
   get_comp_times().then(t => {
-    comp_times = t
+    console.log("got times");
+    comp_times = t;
     if(needs_questions(t)) {
       load_questions();
     }
+    set_timer_for_next_state();
   })
   .catch((err) => {
     loading_error = err
   });
+
 
   const set_solved = (slug: string) => {
     if (questions === undefined) return;
@@ -50,8 +72,8 @@
   }
 </script>
 
-{#if comp_times !== undefined && (!needs_questions(comp_times) || questions !== undefined)}
-  <Client {questions} {set_solved} {comp_times} />
+{#if comp_times !== undefined && current_state !== undefined && (!needs_questions(comp_times) || questions !== undefined)}
+  <Client {questions} {set_solved} {comp_times} {current_state} />
 {:else if loading_error !== undefined}
   <div class="loading">Error loading questions: {loading_error}</div>
 {:else}
